Split map function into separate component
